+++ src/main/resources/db/migration/V5__add_request_auth_client_id.sql
@@ -1,0 +1,1 @@
+ALTER TABLE request ADD COLUMN auth_client_id VARCHAR(255) NOT NULL DEFAULT 'unknown';